Moved PartyDisplay code to separate directory
